pyproject.toml
[project]
name = "ai-business-analysis"
version = "0.1.0"
description = "DAG-based AI business analysis platform"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "celery[redis]>=5.4.0",
    "litellm>=1.55.0",
    "jinja2>=3.1.0",
    "boto3>=1.35.0",
    "httpx>=0.28.0",
    "structlog>=24.4.0",
    "python-multipart>=0.0.18",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "sse-starlette>=2.2.0",
    "tavily-python>=0.5.0",
    "google-search-results>=2.4.0",
    "e2b-code-interpreter>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "celery-types>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
